# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Leave Management System API
  description: "REST API for Leave Management with Approval Chains, Delegations, and\
    \ Escalations"
  contact:
    name: Hackathon Team
    email: team@co.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- BearerAuth: []
...
paths:
  /api/notifications/{id}/read:
    post:
      tags:
      - Notifications
      summary: Mark notification as read
      description: Updates an in-app notification status to read
      operationId: markAsRead
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/manager/requests:
    get:
      tags:
      - Manager
      summary: Get manager approval requests
      description: Retrieves team leave requests assigned to the manager with filtering
        and pagination
      operationId: getManagerRequests
      parameters:
      - name: status
        in: query
        description: Filter by leave status
        required: false
        schema:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
      - name: type
        in: query
        description: Filter by leave type code
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: Filter leaves starting from date
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: Filter leaves ending up to date
        required: false
        schema:
          type: string
          format: date
      - name: q
        in: query
        description: Search query by employee name or reason
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: "Sort expression e.g. createdAt,desc"
        required: false
        schema:
          type: string
          default: "createdAt,desc"
      - name: page
        in: query
        description: Page number (0-indexed)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoLeaveResponseDto"
...
components:
  schemas:
...
    LeaveResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        employeeEmail:
          type: string
        leaveTypeId:
          type: integer
          format: int64
        leaveTypeCode:
          type: string
        leaveTypeName:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        days:
          type: integer
          format: int32
        reason:
          type: string
        status:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
        flagged:
          type: boolean
        flagReason:
          type: string
        createdAt:
          type: string
          format: date-time
        lastActionAt:
          type: string
          format: date-time
...
    PageResponseDtoLeaveResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LeaveResponseDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
...
  securitySchemes:
    BearerAuth:
      type: http
      name: BearerAuth
      scheme: bearer
      bearerFormat: JWT